Fix free/total RAM reporting.
This fixes ramMB for SUSystemProfiler on iOS - was failing to give any data.
This also makes it work if there's ever a device with >4GB of RAM.

For both iOS and OSX we now count inactive pages as free - which should give a
more correct number.

This is related to GH-12, GH-13 and GH-14.