Validate basic sanity of data loaded from disk.
If something is wonky with the data we tried sending, we'll get an error back
and store the data on disk. We'll just keep retrying to send it if the server
continues to reject it each time we start.

This validates some basic sanity of the data we load from disk to prevent this
from happening: Messages without sessions, types or timestamps are discareded on
load.

Resolves GH-24.